Fix #929, #935, #934, #936, #941, #937, #928 and #924

+++ administrator/components/com_kunena/libraries/forum/topic/topic.php
@@ -202,7 +202,7 @@ class KunenaForumTopic extends KunenaDatabaseObject {
 	}
 
 	public function getUserTopic($user=null) {
-		$usertopic = KunenaForumTopicUserHelper::get($this->id, $user);
+		$usertopic = KunenaForumTopicUserHelper::get($this, $user);
 		return $usertopic;
 	}
 
@@ -574,6 +574,8 @@ class KunenaForumTopic extends KunenaDatabaseObject {
 				return false;
 			}
 
+			if ($subjectall) $subject = $target->subject;
+
 		} elseif ($target instanceof KunenaForumCategory) {
 			// Move messages into category
 
@@ -738,15 +740,6 @@ class KunenaForumTopic extends KunenaDatabaseObject {
 			$this->recount();
 			$target->recount();
 		}
-		if ($subject) {
-			// Change subject of first message (if needed)
-			$firstmessage = KunenaForumMessageHelper::get($target->first_post_id, true);
-			if ($firstmessage->subject != $subject) {
-				$firstmessage->subject = $subject;
-				$firstmessage->save();
-			}
-		}
-
 
 		return $target;
 	}
@@ -952,7 +945,7 @@ class KunenaForumTopic extends KunenaDatabaseObject {
 			$this->hold = $exists ? $message->hold : KunenaForum::TOPIC_DELETED;
 			$this->recount();
 		}
-		if (!$this->posts) {
+		if (!($message && $message->exists()) && !$this->posts) {
 			return $this->delete();
 		}
 		if(!$this->save()) {
@@ -1047,7 +1040,7 @@ class KunenaForumTopic extends KunenaDatabaseObject {
 		if (!$this->exists()) {
 			return JText::_ ( 'COM_KUNENA_NO_ACCESS' );
 		}
-		if ($this->hold > 1 || ($this->hold && !$this->getUserTopic($user)->owner)) {
+		if ($this->hold > 1 || ($this->hold)) { // && !$this->getUserTopic($user)->owner)) {
 			$access = KunenaAccess::getInstance();
 			$hold = $access->getAllowedHold($user->userid, $this->category_id, false);
 			if (!in_array($this->hold, $hold)) {
